requirements.txt
langchain
langchain-community
langchain-experimental
python-dotenv
langchain_ollama
reportlab
google-generativeai
fastapi
uvicorn[standard]
